Fix: use setPageMetaInfo function
