+++ packages/evershop/src/modules/cms/pages/admin/cmsPageGrid/index.ts
@@ -1,8 +1,9 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { buildFilterFromUrl } from '../../../../../lib/util/buildFilterFromUrl.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request, response) => {
-  setContextValue(request, 'pageInfo', {
+  setPageMetaInfo(request, {
     title: 'Cms pages',
     description: 'Cms pages'
   });
