+++ packages/evershop/src/modules/cms/pages/admin/cmsPageNew/index.ts
@@ -1,7 +1,7 @@
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 
 export default (request, response) => {
-  setContextValue(request, 'pageInfo', {
+  setPageMetaInfo(request, {
     title: 'Create a new cms page',
     description: 'Create a new cms page'
   });
