--- packages/evershop/src/modules/cms/pages/admin/dashboard/index.js
@@ -1,8 +1,0 @@
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
-
-export default (request, response) => {
-  setContextValue(request, 'pageInfo', {
-    title: 'Dashboard',
-    description: 'dashboard'
-  });
-};
