+++ packages/evershop/src/modules/cms/pages/admin/widgetEdit/index.ts
@@ -2,8 +2,10 @@ import { select } from '@evershop/postgres-query-builder';
 import { pool } from '../../../../../lib/postgres/connection.js';
 import { getEnabledWidgets } from '../../../../../lib/widget/widgetManager.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { setPageMetaInfo } from '../../../services/pageMetaInfo.js';
+import { EvershopResponse } from '../../../../../types/response.js';
 
-export default async (request, response, next) => {
+export default async (request, response: EvershopResponse, next) => {
   try {
     const query = select();
     query.from('widget');
@@ -20,7 +22,7 @@ export default async (request, response, next) => {
       setContextValue(request, 'type', widget.type);
       setContextValue(request, 'widgetId', widget.widget_id);
       setContextValue(request, 'widgetUuid', widget.uuid);
-      setContextValue(request, 'pageInfo', {
+      setPageMetaInfo(request, {
         title: widget.name,
         description: widget.name
       });
