+++ packages/evershop/src/modules/cms/pages/admin/widgetGrid/index.ts
@@ -1,10 +1,11 @@
+import { setPageMetaInfo } from '../../../services/pageMetaInfo.js';
 import { buildFilterFromUrl } from '../../../../../lib/util/buildFilterFromUrl.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request, response) => {
-  setContextValue(request, 'pageInfo', {
-    title: 'Orders',
-    description: 'Orders'
+  setPageMetaInfo(request, {
+    title: 'Widgets',
+    description: 'Widgets'
   });
   setContextValue(request, 'filtersFromUrl', buildFilterFromUrl(request));
 };
