+++ packages/evershop/src/modules/customer/pages/admin/customerEdit/index.ts
@@ -1,8 +1,10 @@
 import { select } from '@evershop/postgres-query-builder';
 import { pool } from '../../../../../lib/postgres/connection.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
+import { EvershopResponse } from '../../../../../types/response.js';
 
-export default async (request, response, next) => {
+export default async (request, response: EvershopResponse, next) => {
   try {
     const query = select();
     query.from('customer');
@@ -15,7 +17,7 @@ export default async (request, response, next) => {
     } else {
       setContextValue(request, 'customerId', customer.customer_id);
       setContextValue(request, 'customerUuid', customer.uuid);
-      setContextValue(request, 'pageInfo', {
+      setPageMetaInfo(request, {
         title: customer.full_name,
         description: customer.full_name
       });
