+++ packages/evershop/src/modules/customer/pages/admin/customerGrid/index.ts
@@ -1,10 +1,11 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { buildFilterFromUrl } from '../../../../../lib/util/buildFilterFromUrl.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request, response) => {
-  setContextValue(request, 'pageInfo', {
-    title: 'Widgets',
-    description: 'Widgets'
+  setPageMetaInfo(request, {
+    title: 'Customers',
+    description: 'Customers'
   });
   setContextValue(request, 'filtersFromUrl', buildFilterFromUrl(request));
 };
