--- packages/evershop/src/modules/customer/pages/frontStore/login/index.js
@@ -1,17 +1,0 @@
-import { translate } from '../../../../../lib/locale/translate/translate.js';
-import { buildUrl } from '../../../../../lib/router/buildUrl.js';
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
-
-export default (request, response, next) => {
-  // Check if the user is logged in
-  if (request.isCustomerLoggedIn()) {
-    // Redirect to homepage
-    response.redirect(buildUrl('homepage'));
-  } else {
-    setContextValue(request, 'pageInfo', {
-      title: translate('Login'),
-      description: translate('Login')
-    });
-    next();
-  }
-};
