+++ packages/evershop/src/modules/customer/pages/frontStore/login/index.ts
@@ -1,16 +1,16 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { translate } from '../../../../../lib/locale/translate/translate.js';
 import { buildUrl } from '../../../../../lib/router/buildUrl.js';
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request, response, next) => {
-  // Check if the customer is logged in
+  // Check if the user is logged in
   if (request.isCustomerLoggedIn()) {
     // Redirect to homepage
     response.redirect(buildUrl('homepage'));
   } else {
-    setContextValue(request, 'pageInfo', {
-      title: translate('Reset password'),
-      description: translate('Reset password')
+    setPageMetaInfo(request, {
+      title: translate('Login'),
+      description: translate('Login')
     });
     next();
   }
