+++ packages/evershop/src/modules/customer/pages/frontStore/register/index.ts
@@ -1,9 +1,7 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { translate } from '../../../../../lib/locale/translate/translate.js';
 import { buildUrl } from '../../../../../lib/router/buildUrl.js';
-import {
-  getContextValue,
-  setContextValue
-} from '../../../../graphql/services/contextHelper.js';
+import { getContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request, response, next) => {
   // Check if the user is logged in
@@ -12,7 +10,7 @@ export default (request, response, next) => {
     // Redirect to admin dashboard
     response.redirect(buildUrl('homepage'));
   } else {
-    setContextValue(request, 'pageInfo', {
+    setPageMetaInfo(request, {
       title: translate('Create an account'),
       description: translate('Create an account')
     });
