+++ packages/evershop/src/modules/customer/pages/frontStore/resetPasswordPage/index.ts
@@ -1,0 +1,19 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
+import { translate } from '../../../../../lib/locale/translate/translate.js';
+import { buildUrl } from '../../../../../lib/router/buildUrl.js';
+import { EvershopRequest } from '../../../../../types/request.js';
+import { EvershopResponse } from '../../../../../types/response.js';
+
+export default (request: EvershopRequest, response: EvershopResponse, next) => {
+  // Check if the customer is logged in
+  if (request.isCustomerLoggedIn()) {
+    // Redirect to homepage
+    response.redirect(buildUrl('homepage'));
+  } else {
+    setPageMetaInfo(request, {
+      title: translate('Reset password'),
+      description: translate('Reset password')
+    });
+    next();
+  }
+};
