+++ packages/evershop/src/modules/customer/pages/frontStore/updatePasswordPage/index.ts
@@ -1,14 +1,16 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { translate } from '../../../../../lib/locale/translate/translate.js';
 import { buildUrl } from '../../../../../lib/router/buildUrl.js';
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { EvershopRequest } from '../../../../../types/request.js';
+import { EvershopResponse } from '../../../../../types/response.js';
 
-export default (request, response, next) => {
+export default (request: EvershopRequest, response: EvershopResponse, next) => {
   // Check if the customer is logged in
   if (request.isCustomerLoggedIn()) {
     // Redirect to homepage
     response.redirect(buildUrl('homepage'));
   } else {
-    setContextValue(request, 'pageInfo', {
+    setPageMetaInfo(request, {
       title: translate('Update password'),
       description: translate('Update password')
     });
