+++ packages/evershop/src/modules/promotion/pages/admin/couponEdit/index.ts
@@ -2,8 +2,10 @@ import { select } from '@evershop/postgres-query-builder';
 import { pool } from '../../../../../lib/postgres/connection.js';
 import { buildUrl } from '../../../../../lib/router/buildUrl.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
+import { EvershopResponse } from '../../../../../types/response.js';
 
-export default async (request, response, next) => {
+export default async (request, response: EvershopResponse, next) => {
   try {
     const query = select();
     query.from('coupon');
@@ -15,7 +17,7 @@ export default async (request, response, next) => {
     } else {
       setContextValue(request, 'couponId', parseInt(coupon.coupon_id, 10));
       setContextValue(request, 'couponUuid', coupon.uuid);
-      setContextValue(request, 'pageInfo', {
+      setPageMetaInfo(request, {
         title: coupon.coupon,
         description: coupon.coupon
       });
