+++ packages/evershop/src/modules/promotion/pages/admin/couponGrid/index.ts
@@ -1,8 +1,9 @@
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 import { buildFilterFromUrl } from '../../../../../lib/util/buildFilterFromUrl.js';
 import { setContextValue } from '../../../../graphql/services/contextHelper.js';
 
 export default (request) => {
-  setContextValue(request, 'pageInfo', {
+  setPageMetaInfo(request, {
     title: 'Coupons',
     description: 'Coupons'
   });
