+++ packages/evershop/src/modules/setting/pages/admin/storeSetting/index.ts
@@ -1,7 +1,7 @@
-import { setContextValue } from '../../../../graphql/services/contextHelper.js';
+import { setPageMetaInfo } from '../../../../cms/services/pageMetaInfo.js';
 
 export default (request) => {
-  setContextValue(request, 'pageInfo', {
+  setPageMetaInfo(request, {
     title: 'Store Setting',
     description: 'Store Setting'
   });
